+++ views/layouts/partials/_sidebar.php
@@ -93,27 +93,10 @@
                         <?php if ($rol_sesion == 'Administrador' || $rol_sesion == 'Comprador') : ?>
                             <!-- Modulo de proveedores -->
                             <li class="nav-item">
-                                <a href="#" class="nav-link active">
+                                <a href="<?= BASE_URL ?>/suppliers" class="nav-link active">
                                     <i class="nav-icon fas fa-truck"></i>
-                                    <p>
-                                        Proveedores
-                                        <i class="right fas fa-angle-left"></i>
-                                    </p>
+                                    <p>Proveedores</p>
                                 </a>
-                                <ul class="nav nav-treeview">
-                                    <li class="nav-item">
-                                        <a href="<?= BASE_URL ?>/suppliers" class="nav-link">
-                                            <i class="far fa-circle nav-icon"></i>
-                                            <p>Lista de proveedores</p>
-                                        </a>
-                                    </li>
-                                    <li class="nav-item">
-                                        <a href="<?= BASE_URL ?>/suppliers/create" class="nav-link">
-                                            <i class="far fa-circle nav-icon"></i>
-                                            <p>Crear proveedor</p>
-                                        </a>
-                                    </li>
-                                </ul>
                             </li>
                             <!-- Modulo de compras -->
                             <li class="nav-item">